+++ driver-core/src/main/java/com/datastax/driver/core/RequestHandler.java
@@ -24,8 +24,9 @@ import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
+import java.util.concurrent.*;
+import java.util.concurrent.locks.Lock;
+import java.util.concurrent.locks.ReentrantLock;
 
 import com.codahale.metrics.Timer;
 import org.slf4j.Logger;
@@ -70,6 +71,10 @@ class RequestHandler implements Connection.ResponseCallback {
 
     private final Timer.Context timerContext;
     private final long startTime;
+
+    // Tracks whether there is a retry already in progress
+    private final Lock retryLock = new ReentrantLock();
+    private Future<?> currentRetry; // guarded by retryLock
 
     public RequestHandler(SessionManager manager, Callback callback, Statement statement) {
         this.manager = manager;
@@ -162,24 +167,34 @@ class RequestHandler implements Connection.ResponseCallback {
     }
 
     private void retry(final boolean retryCurrent, ConsistencyLevel newConsistencyLevel) {
-        final Host h = current;
-        this.retryConsistencyLevel = newConsistencyLevel;
+        try {
+            // We lock to prevent two retries from launching concurrently
+            retryLock.lock();
 
-        // We should not retry on the current thread as this will be an IO thread.
-        manager.executor().execute(new Runnable() {
-            @Override
-            public void run() {
-                try {
-                    if (retryCurrent) {
-                        if (query(h))
-                            return;
+            if (currentRetry != null && !currentRetry.isDone())
+                return;
+
+            final Host h = current;
+            this.retryConsistencyLevel = newConsistencyLevel;
+
+            // We should not retry on the current thread as this will be an IO thread.
+            currentRetry = manager.executor().submit(new Runnable() {
+                @Override
+                public void run() {
+                    try {
+                        if (retryCurrent) {
+                            if (query(h))
+                                return;
+                        }
+                        sendRequest();
+                    } catch (Exception e) {
+                        setFinalException(null, new DriverInternalError("Unexpected exception while retrying query", e));
                     }
-                    sendRequest();
-                } catch (Exception e) {
-                    setFinalException(null, new DriverInternalError("Unexpected exception while retrying query", e));
                 }
-            }
-        });
+            });
+        } finally {
+            retryLock.unlock();
+        }
     }
 
     public void cancel() {
@@ -330,6 +345,16 @@ class RequestHandler implements Connection.ResponseCallback {
                             // Try another node
                             logger.warn("Host {} is overloaded, trying next host.", connection.address);
                             logError(connection.address, new DriverException("Host overloaded"));
+                            if (metricsEnabled())
+                                metrics().getErrorMetrics().getOthers().inc();
+                            retry(false, null);
+                            return;
+                        case SERVER_ERROR:
+                            // Defunct connection and try another node
+                            logger.warn("{} replied with server error ({}), trying next host.", connection.address, err.message);
+                            DriverException exception = new DriverException("Host replied with server error: " + err.message);
+                            logError(connection.address, exception);
+                            connection.defunct(exception);
                             if (metricsEnabled())
                                 metrics().getErrorMetrics().getOthers().inc();
                             retry(false, null);
@@ -498,10 +523,12 @@ class RequestHandler implements Connection.ResponseCallback {
     public void onTimeout(Connection connection, long latency) {
         Host queriedHost = current;
         try {
-            if (connection instanceof PooledConnection)
-                ((PooledConnection)connection).release();
+            // If a query times out, we consider that the host is unstable, so we defunct
+            // the connection to mark it down.
+            DriverException timeoutException = new DriverException("Timed out waiting for server response");
+            connection.defunct(timeoutException);
 
-            logError(connection.address, new DriverException("Timeout during read"));
+            logError(connection.address, timeoutException);
             retry(false, null);
         } catch (Exception e) {
             // This shouldn't happen, but if it does, we want to signal the callback, not let him hang indefinitively